Tests against the JIRA REST API (#126)
* chore: added mockito as dependency

* fix: getting the CVE fail when including the .RELEASE suffix

Since an error raises as not version exists with the  suffix a binary response is being returned

* feat: tests against the JIRA REST service

* chore: removed MockJiraRestClient so the RestClient can be instantiated as needed

* docs: tests now require the JIRA user and password for connectivity tests

* docs: updated the TOC

* fix: removed duplicated parameters

* fix: removed unused import

* fix: the release being cloned was missing the pom.xml file

The release now includes the  file with the upstream information to perform the clone. The release template was also adapted so the version is the same than the BOM

* fix: Added test profiles (quarkus 1.12) so the github tests don't execute integration tests

Since the integration tests are failing ATM they are removed from the tests executed by github action. Had to upgrade quarkus to 1.12 to be able to have test profiles.

* fix: removed the .RELEASE replacement on start-release

* docs: described the different test profiles and how to execute them

* docs: updated the TOC

* chore: updated the issue service test name to better fit the tested actions

* chore: removed unused code

* chore: added mockito as dependency

* fix: getting the CVE fail when including the .RELEASE suffix

Since an error raises as not version exists with the  suffix a binary response is being returned

* feat: tests against the JIRA REST service

* chore: removed MockJiraRestClient so the RestClient can be instantiated as needed

* docs: tests now require the JIRA user and password for connectivity tests

* docs: updated the TOC

* fix: removed duplicated parameters

* fix: removed unused import

* fix: the release being cloned was missing the pom.xml file

The release now includes the  file with the upstream information to perform the clone. The release template was also adapted so the version is the same than the BOM

* fix: Added test profiles (quarkus 1.12) so the github tests don't execute integration tests

Since the integration tests are failing ATM they are removed from the tests executed by github action. Had to upgrade quarkus to 1.12 to be able to have test profiles.

chore: merged main branch with upgrade to quakus 1.13

* fix: removed the .RELEASE replacement on start-release

* docs: described the different test profiles and how to execute them

* docs: updated the TOC

* chore: updated the issue service test name to better fit the tested actions

* chore: removed unused code